extract render/frame responsibility out of GameLoop into a new FrameLoop
- Added CanvasRendererPort, so the frame loop depends on a narrow renderer interface instead of
  InfiniteCanvasRenderer.
- Moved RAF scheduling, animated GIF/video ticking, render-state preparation, renderer calls, perf
  hooks, render-error forwarding, and video-frame tracking into FrameLoop
- Kept GameLoop as the public facade for existing component call sites.
- GameLoop now owns interaction/input state and delegates frame ticking to FrameLoop.
- Exported CanvasRendererPort from the engine barrel at engine/index.ts
